# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Announcement documentation",
        "contact": {
            "email": "profseur@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "http://127.0.0.1:8000/api/announcements/all": {
            "get": {
                "tags": [
                    "INDEX"
                ],
                "summary": "Get List Of announcements",
                "description": "Returns all announcements",
                "operationId": "getAllannouncements/",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "http://127.0.0.1:8000/api/announcements/get/id": {
            "get": {
                "tags": [
                    "SHOW"
                ],
                "summary": "Avoir une annonce",
                "description": "retourne une annonce",
                "operationId": "get Announcements/",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}